added color and finished keyboard interaction

+++ week_03/algo_week_03_01_johnWhitney/src/testApp.h
@@ -21,6 +21,6 @@ class testApp : public ofBaseApp{
 		void gotMessage(ofMessage msg);
     
         vector <Circle> circles;
-        int numCircles;
+        float numCircles;
 		
 };
